fix go.mod so that module can be imported
