+++ eftl/sdk/golang/eftl/examples/kvset/kvset.go
@@ -5,15 +5,16 @@
 // TIBCO Software Inc., Palo Alto, California, USA
 //
 
-// This is an example of a basic eFTL client which removes a key-value pair from a map.
+// This is an example of a basic eFTL client which sets a key-value pair in a map.
 
 package main
 
 import (
 	"log"
 	"os"
+	"time"
 
-	"tibco.com/eftl"
+	"github.com/TIBCOSoftware/TIBCO-Messaging/eftl/sdk/golang/eftl"
 )
 
 var (
@@ -32,7 +33,7 @@ func main() {
 	if len(os.Args) > 1 {
 		url = os.Args[1]
 	}
-        log.Printf("%s : TIBCO eFTL Version: %s\n", os.Args[0], eftl.Version)
+	log.Printf("%s : TIBCO eFTL Version: %s\n", os.Args[0], eftl.Version)
 
 	// set connection options
 	opts := eftl.DefaultOptions()
@@ -52,10 +53,16 @@ func main() {
 	// get a key-value map
 	kv := conn.KVMap(mapName)
 
-	// remove a key-value pair from the map
-	if err := kv.Remove(keyName); err != nil {
-		log.Printf("key-value remove failed: %s\n", err)
+	// create a messsage
+	val := eftl.Message{
+		"text": "sample text",
+		"time": time.Now(),
+	}
+
+	// set the key-value pair in the map
+	if err = kv.Set(keyName, val); err != nil {
+		log.Printf("key-value set failed: %s\n", err)
 	} else {
-		log.Printf("key-value removed\n")
+		log.Printf("key-value set: %s\n", val)
 	}
 }
