+++ eftl/sdk/golang/eftl/go.mod
@@ -1,0 +1,8 @@
+module github.com/TIBCOSoftware/TIBCO-Messaging/eftl/sdk/golang/eftl
+
+go 1.17
+
+require (
+	github.com/gorilla/websocket v1.5.0
+	github.com/rs/zerolog v1.26.1
+)
