--- eftl/sdk/golang/go.mod
@@ -1,11 +1,0 @@
-module eftl
-
-go 1.17
-
-replace tibco.com/eftl => ../eftl
-
-require (
-	github.com/gorilla/websocket v1.5.0
-	github.com/rs/zerolog v1.26.1
-	tibco.com/eftl v0.0.0-00010101000000-000000000000
-)
